Add unit tests with Google Test

+++ include/neotmx.hpp
@@ -45,8 +45,8 @@ namespace tmx {
 } // namespace tmx
 
 struct tmx::Point {
-    double x = 0;
-    double y = 0;
+    float x = 0;
+    float y = 0;
 };
 
 struct tmx::IntPoint {
